implements delete in frontend and home

+++ backend/api/customer/delete.php
@@ -27,6 +27,14 @@
 
         // Validation
 
+        if(!$data){
+            http_response_code(404);
+            echo json_encode([
+                "status" => "404",
+                "message" => "No data",
+            ]);
+            return;
+        }
         $customer->ID = $data->ID;
 
         $err = $customer->delete();
